myapp/src/api/requirements.txt
streamlit
folium==0.12.1
streamlit-folium
gdacs-api
